Dockerize Client and Server Applications (#2)
* feat: Add serve command and update dependencies

This commit adds a new `serve` command to the package.json file, which serves the app from the `dist` folder. It also updates several dependencies including adding "@heroicons/react" and "@tailwindcss/forms", as well as updating "@commitlint/cli" and "@commitlint/config-conventional".

* feat: Add Dockerfile for client

This commit adds a Dockerfile to build and run the client application in a container. The Dockerfile has two stages, one for building the application and another for running it in production mode. It uses Node.js version 14 as the base image, installs dependencies using yarn, copies files into the container, builds the application, and runs it using yarn serve command.

* feat: Add Dockerfile for server

Added a Dockerfile to build the server image with Node.js 14. The Dockerfile sets up the working directory, installs production dependencies, and starts the server using npm start command.

* feat: Add docker-compose.yml file for client and server

This commit adds a new docker-compose.yml file to the project root directory. The file defines two services, 'client' and 'server', each with their own build context and Dockerfile. The 'client' service is set to run on port 5173 and depends on the 'server' service which runs on port 3001.